include files & types comparison

+++ modules/PriorityQueues/include/Utils.h
@@ -5,7 +5,7 @@
 
 using namespace std;
 
-vector<vector<pair<int, int>>> generateGraph(const int& v, const int& e);
+vector<vector<pair<int, int>>> generateGraph(const size_t& v, const size_t& e);
 vector<int> Dijkstra(const vector<vector<pair<int, int>>>& g);
 
 #endif  // RPIORITYQUEUES_INCLUDE_UTILS_H
